back-end/pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = ""
authors = ["GabrielDeAraujoDeSouza <gabriel.souzaaraujo109@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.124.2"
uvicorn = "^0.38.0"
numpy = "^2.3.5"
pillow = "^12.0.0"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
